util file

+++ js/app/test/custom_components_dev.cc.spec.ts
@@ -1,77 +1,47 @@
 import { test, expect } from "@playwright/test";
-import { spawn, spawnSync } from "node:child_process";
-import which from "which";
+import { spawnSync } from "node:child_process";
+import { launch_app_background, kill_process } from "./utils";
 import { join } from "path";
 
 test("gradio cc dev correcty launches and is interactive", async ({ page }) => {
 	test.setTimeout(60 * 1000);
-
-	const install = spawnSync(
+	console.log("cwd", process.cwd());
+	const create = spawnSync(
 		`gradio cc create MyComponent --no-configure-metadata --template SimpleTextbox --overwrite`,
 		{
 			shell: true,
 			stdio: "pipe",
-			cwd: join(process.cwd(), "mycomponent"),
+			cwd: join(process.cwd(), "..", "preview", "test"),
 			env: {
 				...process.env,
 				PYTHONUNBUFFERED: "true"
 			}
 		}
 	);
-	console.log("install stdout", install.stdout.toString());
-	console.log("install stderr", install.stderr.toString());
+	console.log("install stdout", create.stdout.toString());
 
-	const _process = spawn(`gradio cc dev`, {
-		shell: true,
-		stdio: "pipe",
-		cwd: join(process.cwd(), "mycomponent"),
-		env: {
-			...process.env,
-			PYTHONUNBUFFERED: "true"
-		}
-	});
-
-	console.log("Starting gradio cc dev");
-
-	_process.stdout.setEncoding("utf8");
-	_process.stderr.setEncoding("utf8");
-
-	_process.on("exit", () => kill_process(_process));
-	_process.on("close", () => kill_process(_process));
-	_process.on("disconnect", () => kill_process(_process));
-
-	let port;
-
-	function std_out(data) {
-		const _data: string = data.toString();
-		console.log(_data);
-
-		const portRegExp = /:(\d+)/;
-		const match = portRegExp.exec(_data);
-
-		if (match && match[1] && _data.includes("Go here")) {
-			port = parseInt(match[1], 10);
-		}
+	console.log("install stderr", (create?.error || "No errors from gradio cc create").toString());
+	const {port, process: _process} = await launch_app_background(`gradio cc dev`);
+	try {	
+		await page.goto(`http://localhost:${port}`);
+		await page.getByLabel("x").fill("foo");
+		await page.getByRole("button", { name: "Submit" }).click();
+		await expect(page.getByLabel("output")).toHaveValue("foo");
+	} finally {
+		if (_process) kill_process(_process);
+		spawnSync(
+			`rm -rf ${join(process.cwd(), '..', 'preview', 'test', 'mycomponent')}`,
+			{
+				shell: true,
+				stdio: "pipe",
+				env: {
+					...process.env,
+					PYTHONUNBUFFERED: "true"
+				}
+			}
+		);
 	}
 
-	function std_err(data) {
-		const _data: string = data.toString();
-		console.log(_data);
-	}
 
-	_process.stdout.on("data", std_out);
-	_process.stderr.on("data", std_err);
-
-	while (!port) {
-		await new Promise((r) => setTimeout(r, 1000));
-	}
-	await page.goto(`http://localhost:${port}`);
-	await page.getByLabel("x").fill("foo");
-	await page.getByRole("button", { name: "Submit" }).click();
-	await expect(page.getByLabel("output")).toHaveValue("foo");
-	kill_process(_process);
 });
 
-function kill_process(process) {
-	process.kill("SIGKILL");
-}
